Removed problematic code
This conditional was added so that JSON-encoding the results of a parse would result in the correct string when there were no rel values, using an empty object literal instead of an empty list literal. Unfortunately that obviously causes some problems when trying to use the library, e.g. Fatal Errors.