GH Actions: only run PHPCS once
All sniffs from PHPCS are set up to give the same results independently of the PHP version on which PHPCS is run, so there is no need to run it against multiple PHP versions in the matrix.

With this in mind, I'm suggesting to split off the PHPCS check into its own job.

Within that job, I've also set things up in a way that any CS errors will be shown inline in the code view if the build was run for a PR.